projects/probability exchange/backend/requirements_dflow.txt
# DFlow Solana Integration Dependencies

# Core Solana support
base58==2.1.1              # Base58 encoding/decoding for Solana addresses
pynacl==1.5.0              # Cryptographic signature verification (Ed25519)

# HTTP client (if not already installed)
aiohttp>=3.9.0            # Async HTTP client for API calls

# Existing dependencies (ensure these are present)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=23.2.0
python-dotenv>=1.0.0

# Optional: For enhanced Solana integration
# solders>=0.18.0          # Solana Python SDK (optional, for advanced features)
# anchorpy>=0.17.0         # Anchor framework support (optional)
